requirements.txt
# =============================================================================
# Valura AI — Team Lead Assignment — Core Dependencies
# =============================================================================

# --- Web framework + streaming ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.2
httpx>=0.27.0

# --- Validation ---
pydantic>=2.7.0

# --- LLM ---
openai>=1.30.0
google-genai>=1.0.0

# --- Market data ---
yfinance>=0.2.40
numpy>=1.26.0

# --- Utilities ---
python-dotenv>=1.0.0

# --- Testing ---
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0